Extract test for cancellation
