+++ spec/features/order_spec.rb
@@ -63,30 +63,6 @@ RSpec.feature 'Order', type: :feature do
     end
   end
 
-  feature '予約のキャンセル' do
-    scenario '予約者は自分の予約をキャンセルできる' do
-      order_item = create(:order_item, order: order, lunchbox: lunchbox)
-
-      visit order_order_items_path(order)
-      expect(page).to have_text(order_item.customer_name)
-      expect(page).to have_text('予約取り消し')
-
-      click_link('予約取り消し')
-      expect(page).not_to have_text('予約取り消し')
-      expect(page).to have_text("#{order_item.customer_name} さんの #{order_item.lunchbox.name} の予約を取り消しました。")
-    end
-
-    scenario 'Order が締め切られている場合、予約者は自分の予約をキャンセルできない' do
-      order = create(:order, :closed)
-      order_item = create(:order_item, order: order, lunchbox: lunchbox)
-
-      visit order_order_items_path(order)
-
-      expect(page).to have_text(order_item.customer_name)
-      expect(page).not_to have_link('予約取り消し')
-    end
-  end
-
   feature '予約の修正' do
     scenario '予約者は自分の予約を修正できる' do
       other_lunchbox = create(:lunchbox, name: 'sample弁当-上')
